# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "drmeet-api",
    "version": "1.0.0",
    "description": "A simple Appointment Booking System for small clinics, ready for frontend consumption"
  },
  "servers": [
    {
      "url": "https://drmeet-api.onrender.com/"
    },
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
...
  "paths": {
    "/api/appointments": {
      "get": {
        "summary": "Get all appointments",
        "tags": [
          "Appointments"
        ],
        "responses": {
          "200": {
            "description": "A list of all appointments"
          },
          "500": {
            "description": "An error occurred while fetching appointments"
          }
        }
      },
      "post": {
        "summary": "Create a new appointment",
        "tags": [
          "Appointments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "doctor",
                  "patient",
                  "date",
                  "time"
                ],
                "properties": {
                  "doctor": {
                    "type": "string"
                  },
                  "patient": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "confirmed",
                      "cancelled",
                      "completed"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A new appointment created"
          },
          "400": {
            "description": "Failed to create appointment"
          },
          "500": {
            "description": "An error occurred while creating the appointment"
          }
        }
      }
    },
    "/api/appointments/{id}": {
      "get": {
        "summary": "Get an appointment by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the appointment"
          }
        ],
        "tags": [
          "Appointments"
        ],
        "responses": {
          "200": {
            "description": "The appointment with the specified ID"
          },
          "400": {
            "description": "Invalid appointment ID format"
          },
          "404": {
            "description": "Appointment not found"
          },
          "500": {
            "description": "An error occurred while fetching the appointment"
          }
        }
      },
      "put": {
        "summary": "Update an appointment by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the appointment"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "doctor": {
                    "type": "string"
                  },
                  "patient": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "time": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "confirmed",
                      "cancelled",
                      "completed"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Appointments"
        ],
        "responses": {
          "200": {
            "description": "Appointment updated"
          },
          "400": {
            "description": "Invalid appointment ID format"
          },
          "404": {
            "description": "Appointment not found"
          },
          "500": {
            "description": "An error occurred while updating the appointment"
          }
        }
      },
      "delete": {
        "summary": "Deletes an appointment by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the appointment"
          }
        ],
        "tags": [
          "Appointments"
        ],
        "responses": {
          "200": {
            "description": "Appointment was deleted successfully"
          },
          "400": {
            "description": "Invalid appointment ID format"
          },
          "404": {
            "description": "Appointment not found"
          },
          "500": {
            "description": "An error occurred while deleting the appointment"
          }
        }
      }
    },
...
    "/api/doctors": {
      "get": {
        "summary": "Get all doctors",
        "tags": [
          "Doctors"
        ],
        "responses": {
          "200": {
            "description": "A list of all doctors"
          },
          "500": {
            "description": "An error occurred while fetching doctors"
          }
        }
      },
      "post": {
        "summary": "Create a new doctor",
        "tags": [
          "Doctors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "specialization"
                ],
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "specialization": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A new doctor created"
          },
          "400": {
            "description": "Failed to create doctor"
          },
          "500": {
            "description": "An error occurred while creating the doctor"
          }
        }
      }
    },
    "/api/doctors/{id}": {
      "get": {
        "summary": "Get a doctor by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the doctor"
          }
        ],
        "tags": [
          "Doctors"
        ],
        "responses": {
          "200": {
            "description": "The doctor with the specified ID"
          },
          "400": {
            "description": "Invalid doctor ID format"
          },
          "404": {
            "description": "Doctor not found"
          },
          "500": {
            "description": "An error occurred while fetching the doctor"
          }
        }
      },
      "put": {
        "summary": "Update a doctor by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the doctor"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "specialization": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Doctors"
        ],
        "responses": {
          "200": {
            "description": "Doctor updated"
          },
          "400": {
            "description": "Invalid doctor ID format"
          },
          "404": {
            "description": "Doctor not found"
          },
          "500": {
            "description": "An error occurred while updating the doctor"
          }
        }
      },
      "delete": {
        "summary": "Deletes a doctor by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the doctor"
          }
        ],
        "tags": [
          "Doctors"
        ],
        "responses": {
          "200": {
            "description": "Doctor was deleted successfully"
          },
          "400": {
            "description": "Invalid doctor ID format"
          },
          "404": {
            "description": "Doctor not found"
          },
          "500": {
            "description": "An error occurred while deleting the doctor"
          }
        }
      }
    },
...
  }
}